llm_verifier/requirements.txt
# LLM Verifier Dependencies
requests>=2.28.0
python-dotenv>=1.0.0
pyyaml>=6.0
loguru>=0.7.0
